requirements.txt
celery>=4.3.0
Django>=2.2.0
infi.clickhouse-orm>=1.2.0
redis==3.3.11
